requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
certifi==2023.7.22
click==8.1.7
datadispatch==1.0.0
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
inv==0.1.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pydantic==2.2.1
pydantic_core==2.6.1
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tasks==2.8.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
